Refactor switchboard to be more generic and work for multiple configurations. Also update configurations for new state calling
